CCD-2952: Fix intermittent test failures - ReferenceDataRepositoryIT (#1964)
* CCD-2952: Fix intermittent test failures - ReferenceDataRepositoryIT

* CCD-2952: Fix intermittent test failures - ReferenceDataRepositoryIT

* CCD-2952 Fix intermittent failing Reference Data integration tests

* CCD-2952 Temporarily suppress CVE-2016-1000027

* CCD-2952: Fix intermittent test failures - ReferenceDataRepositoryIT

* CCD-2952 Fix intermittent failing Reference Data integration tests

* Dummy commit to trigger build

* Dummy commit to trigger build

* Removed temporary logs
