+++ src/test/java/uk/gov/hmcts/ccd/TestFixtures.java
@@ -157,7 +157,6 @@ public abstract class TestFixtures {
         return mapper.readValue(inputStream, typeReference);
     }
 
-    @SuppressWarnings("SameParameterValue")
     public static String fromFileAsString(final String filePath) {
         StringBuilder json = new StringBuilder();
         try (final InputStreamReader inputStreamReader = new InputStreamReader(getInputStream(filePath));
