--- src/test/resources/mappings/refdata/get_building_locations.json
@@ -1,28 +1,0 @@
-{
-  "id" : "2D4CF82C-57E8-4D5D-B07D-74507EC00675",
-  "request": {
-    "method": "GET",
-    "urlPath": "/refdata/location/building-locations"
-  },
-  "response": {
-    "status": 200,
-    "headers": {
-      "Content-Type": "application/json"
-    },
-    "jsonBody": [
-      {
-        "building_location_id": "123",
-        "building_location_name": "54 TEST ROAD",
-        "building_location_status": "LIVE",
-        "area": "North",
-        "region": "Midlands",
-        "region_id": "1",
-        "cluster_name": "NBC",
-        "cluster_id": "1",
-        "court_finder_url": "",
-        "postcode": "T33ST",
-        "address": "TEST ROAD"
-      }
-    ]
-  }
-}
